requirements.txt
psycopg2-binary
django
djangorestframework
pillow
